test: split token-exchange spec into per-concern suites for parallel stability
Refs: AM-7303
