+++ gravitee-am-test/specs/gateway/token-exchange/token-exchange-delegation.jest.spec.ts
@@ -24,48 +24,14 @@ import {
   TOKEN_EXCHANGE_TEST,
 } from './fixtures/token-exchange-fixture';
 
-setup();
+setup(120000);
 
-let defaultFixture: TokenExchangeFixture;
-let restrictedFixture: TokenExchangeFixture;
 let delegationFixture: TokenExchangeFixture;
 let impersonationOnlyFixture: TokenExchangeFixture;
 let limitedDepthFixture: TokenExchangeFixture;
-let accessTokenOnlyFixture: TokenExchangeFixture;
-let noScopeFixture: TokenExchangeFixture;
-let customClaimsFixture: TokenExchangeFixture;
 let delegationCustomClaimsFixture: TokenExchangeFixture;
 
 beforeAll(async () => {
-  // Setup default fixture with all token types allowed (impersonation only by default)
-  defaultFixture = await setupTokenExchangeFixture();
-
-  // Setup restricted fixture with only ID tokens allowed as subject
-  restrictedFixture = await setupTokenExchangeFixture({
-    domainNamePrefix: 'token-exchange-id-only',
-    domainDescription: 'Token exchange ID only',
-    clientName: 'token-exchange-id-only-client',
-    grantTypes: ['password', 'urn:ietf:params:oauth:grant-type:token-exchange'],
-    scopes: [
-      { scope: 'openid', defaultScope: true },
-      { scope: 'profile', defaultScope: true },
-    ],
-    allowedSubjectTokenTypes: ['urn:ietf:params:oauth:token-type:id_token'],
-  });
-
-  // Setup fixture with only access_token allowed as requested type (ID token not allowed)
-  accessTokenOnlyFixture = await setupTokenExchangeFixture({
-    domainNamePrefix: 'token-exchange-access-only',
-    domainDescription: 'Token exchange access token only',
-    clientName: 'token-exchange-access-only-client',
-    grantTypes: ['password', 'urn:ietf:params:oauth:grant-type:token-exchange'],
-    scopes: [
-      { scope: 'openid', defaultScope: true },
-      { scope: 'profile', defaultScope: true },
-    ],
-    allowedRequestedTokenTypes: ['urn:ietf:params:oauth:token-type:access_token'],
-  });
-
   // Setup delegation fixture with delegation enabled
   delegationFixture = await setupTokenExchangeFixture({
     domainNamePrefix: 'token-exchange-delegation',
@@ -97,27 +63,6 @@ beforeAll(async () => {
     maxDelegationDepth: 1,
   });
 
-  noScopeFixture = await setupTokenExchangeFixture({
-    domainNamePrefix: 'token-exchange-no-scope',
-    domainDescription: 'Token exchange no-scope (subject token no scopes)',
-    clientName: 'token-exchange-no-scope-client',
-    grantTypes: ['password', 'urn:ietf:params:oauth:grant-type:token-exchange'],
-    scopes: [],
-  });
-
-  customClaimsFixture = await setupTokenExchangeFixture({
-    domainNamePrefix: 'token-exchange-custom-claims',
-    domainDescription: 'Token exchange with custom claims',
-    clientName: 'token-exchange-custom-claims-client',
-    tokenCustomClaims: [
-      {
-        claimName: 'am_tx_user_email',
-        claimValue: "{#context.attributes['user'].email}",
-        tokenType: 'ACCESS_TOKEN',
-      },
-    ],
-  });
-
   delegationCustomClaimsFixture = await setupTokenExchangeFixture({
     domainNamePrefix: 'token-exchange-delegation-custom-claims',
     domainDescription: 'Token exchange delegation with custom claims referencing actor_token',
@@ -142,12 +87,6 @@ beforeAll(async () => {
 });
 
 afterAll(async () => {
-  if (defaultFixture) {
-    await defaultFixture.cleanup();
-  }
-  if (restrictedFixture) {
-    await restrictedFixture.cleanup();
-  }
   if (delegationFixture) {
     await delegationFixture.cleanup();
   }
@@ -157,564 +96,9 @@ afterAll(async () => {
   if (limitedDepthFixture) {
     await limitedDepthFixture.cleanup();
   }
-  if (accessTokenOnlyFixture) {
-    await accessTokenOnlyFixture.cleanup();
-  }
-  if (noScopeFixture) {
-    await noScopeFixture.cleanup();
-  }
-  if (customClaimsFixture) {
-    await customClaimsFixture.cleanup();
-  }
   if (delegationCustomClaimsFixture) {
     await delegationCustomClaimsFixture.cleanup();
   }
-});
-
-describe('Token Exchange grant', () => {
-  it('should exchange subject token and keep expiration constraints', async () => {
-    const { oidc, application, basicAuth, obtainSubjectToken } = defaultFixture;
-
-    // Given: subject token has openid, profile, offline_access.
-    const { accessToken: subjectAccessToken, expiresIn: subjectExpiresIn } = await obtainSubjectToken();
-
-    // When: exchange without scope param.
-    const exchangeResponse = await performPost(
-      oidc.token_endpoint,
-      '',
-      `grant_type=urn:ietf:params:oauth:grant-type:token-exchange&subject_token=${subjectAccessToken}&subject_token_type=urn:ietf:params:oauth:token-type:access_token`,
-      {
-        'Content-type': 'application/x-www-form-urlencoded',
-        Authorization: `Basic ${basicAuth}`,
-      },
-    ).expect(200);
-
-    const exchangedToken = exchangeResponse.body;
-    expect(exchangedToken.issued_token_type).toBe('urn:ietf:params:oauth:token-type:access_token');
-    expect(exchangedToken.refresh_token).toBeUndefined();
-    expect(exchangedToken.expires_in).toBeLessThanOrEqual(subjectExpiresIn);
-    expect(exchangedToken.token_type.toLowerCase()).toBe('bearer');
-
-    const decoded = parseJwt(exchangedToken.access_token);
-    expect(decoded.payload['client_id']).toBe(application.settings.oauth.clientId);
-    // Then: all client default scopes are granted when scope is omitted.
-    const issuedScopes = (decoded.payload['scope'] as string)?.split(' ') ?? [];
-    expect(issuedScopes).toHaveLength(2);
-    expect(issuedScopes).toContain('openid');
-    expect(issuedScopes).toContain('profile');
-    expect(issuedScopes).not.toContain('offline_access');
-    expect(exchangedToken.scope).toBeDefined();
-    const responseScopeSet = new Set((exchangedToken.scope as string).split(' '));
-    expect(responseScopeSet.size).toBe(2);
-    expect(responseScopeSet.has('openid')).toBe(true);
-    expect(responseScopeSet.has('profile')).toBe(true);
-    expect(responseScopeSet.has('offline_access')).toBe(false);
-
-    const introspectResponse = await performPost(
-      oidc.introspection_endpoint,
-      '',
-      `token=${exchangedToken.access_token}`,
-      {
-        'Content-type': 'application/x-www-form-urlencoded',
-        Authorization: `Basic ${basicAuth}`,
-      },
-    ).expect(200);
-
-    expect(introspectResponse.body.active).toBe(true);
-    expect(introspectResponse.body.client_id).toBe(application.settings.oauth.clientId);
-    expect(introspectResponse.body.token_type).toBe('bearer');
-  });
-
-  it('should keep gis claim from subject token', async () => {
-    const { oidc, basicAuth, obtainSubjectToken } = defaultFixture;
-
-    const { accessToken: subjectAccessToken } = await obtainSubjectToken();
-    const subjectDecoded = parseJwt(subjectAccessToken);
-    const subjectGis = subjectDecoded.payload['gis'];
-    expect(subjectGis).toBeDefined();
-
-    const exchangeResponse = await performPost(
-      oidc.token_endpoint,
-      '',
-      `grant_type=urn:ietf:params:oauth:grant-type:token-exchange&subject_token=${subjectAccessToken}&subject_token_type=urn:ietf:params:oauth:token-type:access_token`,
-      {
-        'Content-type': 'application/x-www-form-urlencoded',
-        Authorization: `Basic ${basicAuth}`,
-      },
-    ).expect(200);
-
-    const exchangedDecoded = parseJwt(exchangeResponse.body.access_token);
-    expect(exchangedDecoded.payload['gis']).toBe(subjectGis);
-  });
-
-  it('should expose full user profile in context during local token exchange', async () => {
-    const { oidc, basicAuth, obtainSubjectToken, user } = customClaimsFixture;
-
-    const { accessToken: subjectAccessToken } = await obtainSubjectToken();
-
-    const exchangeResponse = await performPost(
-      oidc.token_endpoint,
-      '',
-      `grant_type=urn:ietf:params:oauth:grant-type:token-exchange&subject_token=${subjectAccessToken}&subject_token_type=urn:ietf:params:oauth:token-type:access_token`,
-      {
-        'Content-type': 'application/x-www-form-urlencoded',
-        Authorization: `Basic ${basicAuth}`,
-      },
-    ).expect(200);
-
-    const exchangedDecoded = parseJwt(exchangeResponse.body.access_token);
-    expect(exchangedDecoded.payload['am_tx_user_email']).toBe(user.email);
-  });
-
-  it('should exchange refresh token when allowed', async () => {
-    const { oidc, application, basicAuth, obtainSubjectToken } = defaultFixture;
-
-    const { refreshToken } = await obtainSubjectToken();
-    expect(refreshToken).toBeDefined();
-
-    const response = await performPost(
-      oidc.token_endpoint,
-      '',
-      `grant_type=urn:ietf:params:oauth:grant-type:token-exchange&subject_token=${refreshToken}&subject_token_type=urn:ietf:params:oauth:token-type:refresh_token`,
-      {
-        'Content-type': 'application/x-www-form-urlencoded',
-        Authorization: `Basic ${basicAuth}`,
-      },
-    ).expect(200);
-
-    expect(response.body.issued_token_type).toBe('urn:ietf:params:oauth:token-type:access_token');
-    expect(response.body.access_token).toBeDefined();
-    expect(response.body.refresh_token).toBeUndefined();
-
-    // Introspect the exchanged token to verify it's valid
-    const introspectResponse = await performPost(
-      oidc.introspection_endpoint,
-      '',
-      `token=${response.body.access_token}`,
-      {
-        'Content-type': 'application/x-www-form-urlencoded',
-        Authorization: `Basic ${basicAuth}`,
-      },
-    ).expect(200);
-
-    expect(introspectResponse.body.active).toBe(true);
-    expect(introspectResponse.body.client_id).toBe(application.settings.oauth.clientId);
-    expect(introspectResponse.body.token_type).toBe('bearer');
-  });
-
-  it('should exchange id token when allowed', async () => {
-    const { oidc, application, basicAuth, obtainSubjectToken } = defaultFixture;
-
-    const { idToken } = await obtainSubjectToken();
-    expect(idToken).toBeDefined();
-
-    const response = await performPost(
-      oidc.token_endpoint,
-      '',
-      `grant_type=urn:ietf:params:oauth:grant-type:token-exchange&subject_token=${idToken}&subject_token_type=urn:ietf:params:oauth:token-type:id_token`,
-      {
-        'Content-type': 'application/x-www-form-urlencoded',
-        Authorization: `Basic ${basicAuth}`,
-      },
-    ).expect(200);
-
-    expect(response.body.issued_token_type).toBe('urn:ietf:params:oauth:token-type:access_token');
-    expect(response.body.access_token).toBeDefined();
-    expect(response.body.refresh_token).toBeUndefined();
-
-    // Introspect the exchanged token to verify it's valid
-    const introspectResponse = await performPost(
-      oidc.introspection_endpoint,
-      '',
-      `token=${response.body.access_token}`,
-      {
-        'Content-type': 'application/x-www-form-urlencoded',
-        Authorization: `Basic ${basicAuth}`,
-      },
-    ).expect(200);
-
-    expect(introspectResponse.body.active).toBe(true);
-    expect(introspectResponse.body.client_id).toBe(application.settings.oauth.clientId);
-    expect(introspectResponse.body.token_type).toBe('bearer');
-  });
-
-  it('should reject when requested_token_type is not supported', async () => {
-    const { oidc, basicAuth, obtainSubjectToken } = defaultFixture;
-
-    const { accessToken: subjectAccessToken } = await obtainSubjectToken('openid%20profile');
-
-    const response = await performPost(
-      oidc.token_endpoint,
-      '',
-      `grant_type=urn:ietf:params:oauth:grant-type:token-exchange&subject_token=${subjectAccessToken}&subject_token_type=urn:ietf:params:oauth:token-type:access_token&requested_token_type=urn:ietf:params:oauth:token-type:refresh_token`,
-      {
-        'Content-type': 'application/x-www-form-urlencoded',
-        Authorization: `Basic ${basicAuth}`,
-      },
-    ).expect(400);
-
-    expect(response.body.error).toBe('invalid_request');
-    expect(response.body.error_description).toContain('requested_token_type');
-  });
-
-  it('should NOT include id_token in response when exchanging access token with openid scope', async () => {
-    const { oidc, basicAuth, obtainSubjectToken } = defaultFixture;
-
-    // Obtain a subject token with openid scope
-    const { accessToken: subjectAccessToken } = await obtainSubjectToken('openid%20profile');
-
-    // Exchange the token without specifying requested_token_type (defaults to access_token)
-    const exchangeResponse = await performPost(
-      oidc.token_endpoint,
-      '',
-      `grant_type=urn:ietf:params:oauth:grant-type:token-exchange&subject_token=${subjectAccessToken}&subject_token_type=urn:ietf:params:oauth:token-type:access_token`,
-      {
-        'Content-type': 'application/x-www-form-urlencoded',
-        Authorization: `Basic ${basicAuth}`,
-      },
-    ).expect(200);
-
-    // Per RFC 8693, token exchange response should NOT include id_token unless explicitly requested
-    expect(exchangeResponse.body.access_token).toBeDefined();
-    expect(exchangeResponse.body.issued_token_type).toBe('urn:ietf:params:oauth:token-type:access_token');
-    expect(exchangeResponse.body.id_token).toBeUndefined();
-  });
-
-  it('should include granted scopes in response when client requests subset', async () => {
-    const { oidc, basicAuth, obtainSubjectToken } = defaultFixture;
-
-    // Given: subject token has openid, profile, offline_access.
-    const { accessToken: subjectAccessToken } = await obtainSubjectToken('openid%20profile%20offline_access');
-
-    // When: exchange with scope=openid.
-    const exchangeResponse = await performPost(
-      oidc.token_endpoint,
-      '',
-      `grant_type=urn:ietf:params:oauth:grant-type:token-exchange&subject_token=${subjectAccessToken}&subject_token_type=urn:ietf:params:oauth:token-type:access_token&scope=openid`,
-      {
-        'Content-type': 'application/x-www-form-urlencoded',
-        Authorization: `Basic ${basicAuth}`,
-      },
-    ).expect(200);
-
-    // Then: response and JWT have scope openid.
-    expect(exchangeResponse.body.access_token).toBeDefined();
-    expect(exchangeResponse.body.scope).toBe('openid');
-    const decoded = parseJwt(exchangeResponse.body.access_token);
-    expect(decoded.payload['scope']).toBe('openid');
-  });
-
-  it('should issue new token with no scopes when subject token has no scopes', async () => {
-    const { oidc, basicAuth, obtainSubjectToken } = noScopeFixture;
-
-    // Given: subject token with no scopes.
-    const { accessToken: subjectAccessToken } = await obtainSubjectToken('');
-
-    const exchangeResponse = await performPost(
-      oidc.token_endpoint,
-      '',
-      `grant_type=urn:ietf:params:oauth:grant-type:token-exchange&subject_token=${subjectAccessToken}&subject_token_type=urn:ietf:params:oauth:token-type:access_token`,
-      {
-        'Content-type': 'application/x-www-form-urlencoded',
-        Authorization: `Basic ${basicAuth}`,
-      },
-    ).expect(200);
-
-    expect(exchangeResponse.body.access_token).toBeDefined();
-    expect(exchangeResponse.body.scope).toBeUndefined();
-    const decoded = parseJwt(exchangeResponse.body.access_token);
-    expect(decoded.payload['scope']).toBeUndefined();
-  });
-
-  it('should return 400 invalid_scope when requested scope is not a subset of subject token scopes', async () => {
-    const { oidc, basicAuth, obtainSubjectToken } = defaultFixture;
-
-    // Given: subject token has openid, profile only.
-    const { accessToken: subjectAccessToken } = await obtainSubjectToken('openid%20profile');
-
-    // When: request scope=offline_access (not in subject).
-    const response = await performPost(
-      oidc.token_endpoint,
-      '',
-      `grant_type=urn:ietf:params:oauth:grant-type:token-exchange&subject_token=${subjectAccessToken}&subject_token_type=urn:ietf:params:oauth:token-type:access_token&scope=offline_access`,
-      {
-        'Content-type': 'application/x-www-form-urlencoded',
-        Authorization: `Basic ${basicAuth}`,
-      },
-    ).expect(400);
-
-    // Then: 400 invalid_scope.
-    expect(response.body.error).toBe('invalid_scope');
-    expect(response.body.error_description).toBeDefined();
-    expect(String(response.body.error_description).toLowerCase()).toMatch(/scope|disallowed|invalid/);
-  });
-
-  it('should exchange jwt subject token when allowed', async () => {
-    const { oidc, application, basicAuth, obtainSubjectToken } = defaultFixture;
-    const { accessToken: subjectAccessToken } = await obtainSubjectToken();
-
-    const response = await performPost(
-      oidc.token_endpoint,
-      '',
-      `grant_type=urn:ietf:params:oauth:grant-type:token-exchange&subject_token=${subjectAccessToken}&subject_token_type=urn:ietf:params:oauth:token-type:jwt`,
-      {
-        'Content-type': 'application/x-www-form-urlencoded',
-        Authorization: `Basic ${basicAuth}`,
-      },
-    ).expect(200);
-
-    expect(response.body.issued_token_type).toBe('urn:ietf:params:oauth:token-type:access_token');
-    expect(response.body.access_token).toBeDefined();
-
-    const decoded = parseJwt(response.body.access_token);
-    expect(decoded.payload['client_id']).toBe(application.settings.oauth.clientId);
-  });
-
-  it('should reject when subject_token_type is not allowed', async () => {
-    const { oidc, basicAuth, obtainSubjectToken } = restrictedFixture;
-
-    const { accessToken: subjectAccessToken } = await obtainSubjectToken('openid%20profile');
-
-    const response = await performPost(
-      oidc.token_endpoint,
-      '',
-      `grant_type=urn:ietf:params:oauth:grant-type:token-exchange&subject_token=${subjectAccessToken}&subject_token_type=urn:ietf:params:oauth:token-type:saml1`,
-      {
-        'Content-type': 'application/x-www-form-urlencoded',
-        Authorization: `Basic ${basicAuth}`,
-      },
-    ).expect(400);
-
-    expect(response.body.error).toBe('invalid_request');
-    expect(response.body.error_description).toContain('subject_token_type');
-  });
-});
-
-describe('Token Exchange with invalid tokens', () => {
-  it('should reject a subject token with a tampered signature', async () => {
-    const { oidc, basicAuth, obtainSubjectToken } = defaultFixture;
-
-    // Given: obtain a valid subject token, then corrupt its signature.
-    const { accessToken: validToken } = await obtainSubjectToken('openid%20profile');
-    const parts = validToken.split('.');
-    const tamperedToken = `${parts[0]}.${parts[1]}.invalidsignatureXXXXXXXXXXX`;
-
-    // When: present the tampered token for exchange.
-    const response = await performPost(
-      oidc.token_endpoint,
-      '',
-      `grant_type=urn:ietf:params:oauth:grant-type:token-exchange&subject_token=${tamperedToken}&subject_token_type=urn:ietf:params:oauth:token-type:access_token`,
-      {
-        'Content-type': 'application/x-www-form-urlencoded',
-        Authorization: `Basic ${basicAuth}`,
-      },
-    ).expect(400);
-
-    // Then: invalid_request with a generic description.
-    expect(response.body.error).toBe('invalid_request');
-    expect(response.body.error_description).toBe('The presented token is invalid');
-  });
-
-  it('should reject a tampered actor token in delegation', async () => {
-    const { oidc, basicAuth, obtainSubjectToken, obtainActorToken } = delegationFixture;
-
-    // Given: valid subject and actor tokens; tamper the actor token's signature.
-    const { accessToken: subjectToken } = await obtainSubjectToken('openid%20profile');
-    const { accessToken: validActorToken } = await obtainActorToken('openid%20profile');
-    const parts = validActorToken.split('.');
-    const tamperedActorToken = `${parts[0]}.${parts[1]}.invalidsignatureXXXXXXXXXXX`;
-
-    // When: present the tampered actor token for delegation exchange.
-    const response = await performPost(
-      oidc.token_endpoint,
-      '',
-      `grant_type=urn:ietf:params:oauth:grant-type:token-exchange` +
-      `&subject_token=${subjectToken}` +
-      `&subject_token_type=urn:ietf:params:oauth:token-type:access_token` +
-      `&actor_token=${tamperedActorToken}` +
-      `&actor_token_type=urn:ietf:params:oauth:token-type:access_token`,
-      {
-        'Content-type': 'application/x-www-form-urlencoded',
-        Authorization: `Basic ${basicAuth}`,
-      },
-    ).expect(400);
-
-    // Then: invalid_request with a generic description.
-    expect(response.body.error).toBe('invalid_request');
-    expect(response.body.error_description).toBe('The presented token is invalid');
-  });
-});
-
-describe('Token Exchange with restricted subject token types', () => {
-  it('should reject access token when domain only allows id tokens', async () => {
-    const { oidc, basicAuth, obtainSubjectToken } = restrictedFixture;
-
-    const { accessToken: subjectAccessToken } = await obtainSubjectToken('openid%20profile');
-
-    const response = await performPost(
-      oidc.token_endpoint,
-      '',
-      `grant_type=urn:ietf:params:oauth:grant-type:token-exchange&subject_token=${subjectAccessToken}&subject_token_type=urn:ietf:params:oauth:token-type:access_token`,
-      {
-        'Content-type': 'application/x-www-form-urlencoded',
-        Authorization: `Basic ${basicAuth}`,
-      },
-    ).expect(400);
-
-    expect(response.body.error).toBe('invalid_request');
-    expect(response.body.error_description).toContain('subject_token_type');
-  });
-
-  it('should accept id token when domain only allows id tokens', async () => {
-    const { oidc, application, basicAuth, obtainSubjectToken } = restrictedFixture;
-
-    const { idToken } = await obtainSubjectToken('openid%20profile');
-    expect(idToken).toBeDefined();
-
-    const response = await performPost(
-      oidc.token_endpoint,
-      '',
-      `grant_type=urn:ietf:params:oauth:grant-type:token-exchange&subject_token=${idToken}&subject_token_type=urn:ietf:params:oauth:token-type:id_token`,
-      {
-        'Content-type': 'application/x-www-form-urlencoded',
-        Authorization: `Basic ${basicAuth}`,
-      },
-    ).expect(200);
-
-    expect(response.body.issued_token_type).toBe('urn:ietf:params:oauth:token-type:access_token');
-    expect(response.body.access_token).toBeDefined();
-
-    // Verify the exchanged token
-    const decoded = parseJwt(response.body.access_token);
-    expect(decoded.payload['client_id']).toBe(application.settings.oauth.clientId);
-  });
-});
-
-describe('Token Exchange with requested_token_type=id_token', () => {
-  it('should return ID token when requested_token_type is id_token', async () => {
-    const { oidc, basicAuth, obtainSubjectToken } = defaultFixture;
-
-    const { accessToken: subjectAccessToken } = await obtainSubjectToken('openid%20profile');
-
-    const response = await performPost(
-      oidc.token_endpoint,
-      '',
-      `grant_type=urn:ietf:params:oauth:grant-type:token-exchange&subject_token=${subjectAccessToken}&subject_token_type=urn:ietf:params:oauth:token-type:access_token&requested_token_type=urn:ietf:params:oauth:token-type:id_token`,
-      {
-        'Content-type': 'application/x-www-form-urlencoded',
-        Authorization: `Basic ${basicAuth}`,
-      },
-    ).expect(200);
-
-    // Per RFC 8693, when ID token is requested:
-    // - issued_token_type should be urn:ietf:params:oauth:token-type:id_token
-    // - token_type should be "N_A" (not applicable)
-    // - access_token field contains the ID token
-    expect(response.body.issued_token_type).toBe('urn:ietf:params:oauth:token-type:id_token');
-    expect(response.body.token_type).toBe('N_A');
-    expect(response.body.access_token).toBeDefined();
-    expect(response.body.refresh_token).toBeUndefined();
-
-    // Verify it's a valid JWT and contains expected ID token claims
-    const decoded = parseJwt(response.body.access_token);
-    expect(decoded.payload['sub']).toBeDefined();
-    expect(decoded.payload['iss']).toBeDefined();
-    expect(decoded.payload['aud']).toBeDefined();
-    expect(decoded.payload['exp']).toBeDefined();
-    expect(decoded.payload['iat']).toBeDefined();
-  });
-
-  it('should return ID token with expires_in consistent with subject token expiration', async () => {
-    const { oidc, basicAuth, obtainSubjectToken } = defaultFixture;
-
-    const { accessToken: subjectAccessToken } = await obtainSubjectToken('openid%20profile');
-
-    // Decode the subject token to get its expiration time
-    const subjectPayload = JSON.parse(Buffer.from(subjectAccessToken.split('.')[1], 'base64').toString());
-    const subjectExp = subjectPayload.exp;
-    const nowSeconds = Math.floor(Date.now() / 1000);
-    const subjectRemainingTime = subjectExp - nowSeconds;
-
-    const response = await performPost(
-      oidc.token_endpoint,
-      '',
-      `grant_type=urn:ietf:params:oauth:grant-type:token-exchange&subject_token=${subjectAccessToken}&subject_token_type=urn:ietf:params:oauth:token-type:access_token&requested_token_type=urn:ietf:params:oauth:token-type:id_token`,
-      {
-        'Content-type': 'application/x-www-form-urlencoded',
-        Authorization: `Basic ${basicAuth}`,
-      },
-    ).expect(200);
-
-    // expires_in should be present and reasonable
-    expect(response.body.expires_in).toBeDefined();
-    expect(response.body.expires_in).toBeGreaterThan(0);
-
-    // The exchanged token's expires_in should not exceed the subject token's remaining time
-    // Allow a small buffer (5 seconds) for processing time
-    expect(response.body.expires_in).toBeLessThanOrEqual(subjectRemainingTime + 5);
-  });
-
-  it('should NOT include scope in ID token response', async () => {
-    const { oidc, basicAuth, obtainSubjectToken } = defaultFixture;
-
-    // Request with openid scope - but ID token response should not include scope
-    const { accessToken: subjectAccessToken } = await obtainSubjectToken('openid%20profile');
-
-    const response = await performPost(
-      oidc.token_endpoint,
-      '',
-      `grant_type=urn:ietf:params:oauth:grant-type:token-exchange&subject_token=${subjectAccessToken}&subject_token_type=urn:ietf:params:oauth:token-type:access_token&requested_token_type=urn:ietf:params:oauth:token-type:id_token`,
-      {
-        'Content-type': 'application/x-www-form-urlencoded',
-        Authorization: `Basic ${basicAuth}`,
-      },
-    ).expect(200);
-
-    // ID tokens are for identity/authentication, not authorization
-    // Therefore scope should NOT be included in the response
-    expect(response.body.scope).toBeUndefined();
-    expect(response.body.issued_token_type).toBe('urn:ietf:params:oauth:token-type:id_token');
-  });
-
-  it('should reject id_token request when not in allowedRequestedTokenTypes', async () => {
-    const { oidc, basicAuth, obtainSubjectToken } = accessTokenOnlyFixture;
-
-    const { accessToken: subjectAccessToken } = await obtainSubjectToken('openid%20profile');
-
-    const response = await performPost(
-      oidc.token_endpoint,
-      '',
-      `grant_type=urn:ietf:params:oauth:grant-type:token-exchange&subject_token=${subjectAccessToken}&subject_token_type=urn:ietf:params:oauth:token-type:access_token&requested_token_type=urn:ietf:params:oauth:token-type:id_token`,
-      {
-        'Content-type': 'application/x-www-form-urlencoded',
-        Authorization: `Basic ${basicAuth}`,
-      },
-    ).expect(400);
-
-    expect(response.body.error).toBe('invalid_request');
-    expect(response.body.error_description).toContain('requested_token_type');
-  });
-
-  it('should use access_token as default when not specified', async () => {
-    const { oidc, basicAuth, obtainSubjectToken } = defaultFixture;
-
-    const { accessToken: subjectAccessToken } = await obtainSubjectToken('openid%20profile');
-
-    // Request without requested_token_type should default to access_token
-    const response = await performPost(
-      oidc.token_endpoint,
-      '',
-      `grant_type=urn:ietf:params:oauth:grant-type:token-exchange&subject_token=${subjectAccessToken}&subject_token_type=urn:ietf:params:oauth:token-type:access_token`,
-      {
-        'Content-type': 'application/x-www-form-urlencoded',
-        Authorization: `Basic ${basicAuth}`,
-      },
-    ).expect(200);
-
-    // Should default to access_token
-    expect(response.body.issued_token_type).toBe('urn:ietf:params:oauth:token-type:access_token');
-    expect(response.body.token_type.toLowerCase()).toBe('bearer');
-  });
 });
 
 describe('Token Exchange Delegation (RFC 8693)', () => {
